+++ src/001-050/S45.cc
@@ -7,21 +7,21 @@ namespace ProjectEuler
 {
 std::string S45::GetAnswer()
 {
-	std::vector<long long> Ts(N);
-	std::vector<long long> Ps(N);
-	std::vector<long long> Hs(N);
+	std::vector<int64_t> Ts(N);
+	std::vector<int64_t> Ps(N);
+	std::vector<int64_t> Hs(N);
 
-	for (long long i = 1; i <= N; i++)
+	for (int64_t i = 1; i <= N; i++)
 	{
 		Ts[i - 1] = i * (i + 1) / 2;
 		Ps[i - 1] = i * (3 * i - 1) / 2;
 		Hs[i - 1] = i * (2 * i - 1);
 	}
 
-	std::vector<long long> tp_intersection;
+	std::vector<int64_t> tp_intersection;
 	std::set_intersection(Ts.begin(), Ts.end(), Ps.begin(), Ps.end(), std::back_inserter(tp_intersection));
 
-	std::vector<long long> final_intersection;
+	std::vector<int64_t> final_intersection;
 	std::set_intersection(tp_intersection.begin(),
 						  tp_intersection.end(),
 						  Hs.begin(),
