+++ src/051-100/S70.cc
@@ -9,7 +9,7 @@ using namespace std;
 
 namespace ProjectEuler
 {
-bool IsPermutation(long long a, long long b)
+bool IsPermutation(int64_t a, int64_t b)
 {
 	string sa = to_string(a);
 	string sb = to_string(b);
@@ -24,13 +24,13 @@ bool IsPermutation(long long a, long long b)
 
 std::string S70::GetAnswer()
 {
-	long long limit = 10000000;
+	int64_t limit = 10000000;
 	double minRatio = 100.0;
-	long long result = 0;
+	int64_t result = 0;
 
 	auto primes = Utils::GenPrimes(5000);
 	std::vector<int> rangePrimes;
-	for (long long p : primes)
+	for (int64_t p : primes)
 	{
 		if (p > 2000)
 		{
@@ -42,13 +42,13 @@ std::string S70::GetAnswer()
 	{
 		for (size_t j = i; j < rangePrimes.size(); j++)
 		{
-			long long n = (long long)rangePrimes[i] * rangePrimes[j];
+			int64_t n = (int64_t)rangePrimes[i] * rangePrimes[j];
 			if (n > limit)
 			{
 				break;
 			}
 
-			long long phi = (rangePrimes[i] - 1) * (rangePrimes[j] - 1);
+			int64_t phi = (rangePrimes[i] - 1) * (rangePrimes[j] - 1);
 
 			double ratio = (double)n / phi;
 			if (ratio < minRatio)
