+++ src/051-100/S71.cc
@@ -9,9 +9,9 @@ namespace ProjectEuler
 {
 std::string S71::GetAnswer()
 {
-	long long bestN = 0;
-	long long bestD = 1;
-	long long limit = 1000000;
+	int64_t bestN = 0;
+	int64_t bestD = 1;
+	int64_t limit = 1000000;
 
 	// We want n/d just less than 3/7
 	// n/d < 3/7  =>  7n < 3d
@@ -24,14 +24,14 @@ std::string S71::GetAnswer()
 	// n = floor((3d - 1) / 7).
 	// Compare n/d with bestN/bestD.
 
-	for (long long d = 1; d <= limit; d++)
+	for (int64_t d = 1; d <= limit; d++)
 	{
 		if (d == 7)
 		{
 			continue; // 3/7 itself not allowed? "strictly less"
 		}
 
-		long long n = (3 * d - 1) / 7;
+		int64_t n = (3 * d - 1) / 7;
 		if (n * bestD > bestN * d)
 		{
 			bestN = n;
