+++ src/051-100/S75.cc
@@ -17,16 +17,16 @@ std::string S75::GetAnswer()
 	// gcd(m, n) == 1
 	// m, n not both odd (one even, one odd)
 
-	long long m_limit = sqrt(limit / 2);
+	int64_t m_limit = sqrt(limit / 2);
 
-	for (long long m = 2; m <= m_limit; m++)
+	for (int64_t m = 2; m <= m_limit; m++)
 	{
-		for (long long n = 1; n < m; n++)
+		for (int64_t n = 1; n < m; n++)
 		{
 			if ((m + n) % 2 == 1 && Utils::GetGcd(m, n) == 1)
 			{
-				long long L = 2 * m * (m + n);
-				long long k = 1;
+				int64_t L = 2 * m * (m + n);
+				int64_t k = 1;
 				while (k * L <= limit)
 				{
 					counts[k * L]++;
