+++ src/101-150/S108.h
@@ -10,10 +10,10 @@ class S108 : public ISolution
 {
 public:
 	std::string GetAnswer() override;
-	void Search(int index, long long current_n, long long count, int limit, int max_exponent);
+	void Search(int index, int64_t current_n, int64_t count, int limit, int max_exponent);
 
 private:
 	std::vector<int> primes = { 2, 3, 5, 7, 11, 13, 17, 19, 23, 29, 31, 37, 41, 43, 47 };
-	long long min_n = LLONG_MAX;
+	int64_t min_n = LLONG_MAX;
 };
 } // namespace ProjectEuler
