+++ src/101-150/S117.cc
@@ -9,9 +9,9 @@ namespace ProjectEuler
 {
 std::string S117::GetAnswer()
 {
-	vector<vector<long long>> reds(3, vector<long long>(Layer, 0));
-	vector<vector<long long>> greens(4, vector<long long>(Layer, 0));
-	vector<vector<long long>> blues(5, vector<long long>(Layer, 0));
+	vector<vector<int64_t>> reds(3, vector<int64_t>(Layer, 0));
+	vector<vector<int64_t>> greens(4, vector<int64_t>(Layer, 0));
+	vector<vector<int64_t>> blues(5, vector<int64_t>(Layer, 0));
 
 	reds[0][0] = 1;
 	reds[1][0] = 1;
@@ -28,7 +28,7 @@ std::string S117::GetAnswer()
 
 	for (int i = 1; i < Layer; i++)
 	{
-		long long next_blank = reds[0][i - 1] + reds[2][i - 1] + greens[3][i - 1] + blues[4][i - 1];
+		int64_t next_blank = reds[0][i - 1] + reds[2][i - 1] + greens[3][i - 1] + blues[4][i - 1];
 		reds[1][i] = reds[0][i] = next_blank;
 		greens[1][i] = greens[0][i] = next_blank;
 		blues[1][i] = blues[0][i] = next_blank;
@@ -41,7 +41,7 @@ std::string S117::GetAnswer()
 		blues[4][i] = blues[3][i - 1];
 	}
 
-	long long result = reds[0][Layer - 1] + reds[2][Layer - 1] + greens[3][Layer - 1] + blues[4][Layer - 1];
+	int64_t result = reds[0][Layer - 1] + reds[2][Layer - 1] + greens[3][Layer - 1] + blues[4][Layer - 1];
 	return std::to_string(result);
 }
